+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.9.0 (unreleased)
+# 0.9.0 (May 15, 2021)
 
 ### Changed